[e2e] move flags to common
